RAG_QnA_for_Health_Supplements/requirements.txt
streamlit
python-dotenv
langchain_ollama
faiss-cpu
langchain-community
langchain-core

